PO-1993 resolving comment

+++ src/main/java/uk/gov/hmcts/opal/service/MinorCreditorService.java
@@ -111,7 +111,7 @@ public class MinorCreditorService {
         UserState userState = userStateService.checkForAuthorisedUser(authHeaderValue);
         if (businessUnitId == null) {
             throw new PermissionNotAllowedException(
-                null,
+                (Short) null,
                 FinesPermission.ADD_AND_REMOVE_PAYMENT_HOLD);
         }
         Short businessUnitIdShort = Short.valueOf(businessUnitId);
